app/requirements.txt
streamlit==1.28.1
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow
numpy
pandas==2.0.3
plotly==5.17.0
pytube==15.0.0
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2
altair==5.1.2
streamlit-option-menu==0.3.6
streamlit-extras==0.3.5
